# api-backend/openapi.yaml
openapi: 3.0.3

info:
  title: Todo List API
  description: |
    Simple REST API for managing personal todo lists with JWT-based authentication.
    Built for Pre-test Assignment.
  version: 1.0.0
  contact:
    name: Ace Of Aces
    # url: https://github.com/yourusername/todo-api-backend
  license:
    name: ISC

servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://your-project-name.vercel.app
    description: Production server (Vercel)

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT token in format: Bearer {token}"

  schemas:
    UserRegisterInput:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
          minLength: 3
          example: johndoe
        email:
          type: string
          format: email
          example: john.doe@example.com
        password:
          type: string
          minLength: 6
          example: secret123

    UserLoginInput:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: john.doe@example.com
        password:
          type: string
          example: secret123

    TodoInput:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          example: Buy groceries
        description:
          type: string
          example: Milk, bread, eggs
        completed:
          type: boolean
          default: false
          example: false

    Todo:
      type: object
      properties:
        id:
          type: integer
          example: 1
        userId:
          type: integer
          example: 5
        title:
          type: string
          example: Buy groceries
        description:
          type: string
          nullable: true
          example: Milk, bread, eggs
        completed:
          type: boolean
          example: false
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true

    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Invalid token
        errors:
          type: array
          items:
            type: object
          nullable: true

  responses:
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

paths:
  /:
    get:
      summary: API Information
      description: Returns basic information about the API and available endpoints
      tags:
        - General
      responses:
        '200':
          description: API welcome message
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  version:
                    type: string
                  endpoints:
                    type: object

  /api/auth/register:
    post:
      summary: Register a new user
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegisterInput'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      user:
                        type: object
                        properties:
                          id: { type: integer }
                          username: { type: string }
                          email: { type: string }
                      token:
                        type: string
        '400':
          description: Validation error or user already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'



  /api/todos/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
        description: Todo ID

    get:
      summary: Get a single todo by ID
      tags:
        - Todos
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Todo found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    $ref: '#/components/schemas/Todo'
        '404':
          description: Todo not found
        '401':
          $ref: '#/components/responses/Unauthorized'

  